fix(types): Add type annotations and assertions (700 → 661 errors)
Type fixes applied:
- Add assertions after lazy loading for type narrowing
- Add float() cast for numeric return types
- Add Dict/List type annotations for empty collections
- Use Optional for dataclass fields with None defaults
- Fix variable shadowing in loops with different types
- Add missing Any imports

+++ src/beanllm/infrastructure/routing/routing_rules.py
@@ -154,4 +154,4 @@ class CompositeRule(RoutingRule):
 
         weighted_sum = sum(rule.evaluate(model, request) * weight for rule, weight in self.rules)
 
-        return weighted_sum / total_weight
+        return float(weighted_sum / total_weight)
